requirements.txt
# Core Libraries
pandas
numpy

# Data Ingestion
yfinance
fredapi
feedparser
beautifulsoup4 # For cleaning HTML/XML from RSS

# RAG & Vector Database
chromadb
sentence-transformers # For Hugging Face embedding models
llama-index # High-level RAG orchestration
ollama # To connect to local LLMs (Gemma, Llama 3.1)

# Machine Learning
scikit-learn
lightgbm
xgboost

# Dashboard & App
streamlit
pyyaml # For loading the config.yaml file
fastapi
uvicorn
google-generativeai
pandas-ta-openbb
importlib-metadata
qdrant-client
faiss-cpu
sentence-transformers
hmmlearn
prefect